refactor - unify styles and remove duplicate code
