+++ client/app/(pages)/(user)/schedule.tsx
@@ -1,12 +1,10 @@
 import { useAuth } from "@/lib/auth";
 import { useTheme } from "@/lib/bootstrap/ThemeProvider";
 import { ScheduleList } from "@/lib/components/Schedule";
+import { inset } from "@/lib/theme/spacing";
 import { router } from "expo-router";
 import { useEffect, useMemo } from "react";
-import {
-  StyleSheet,
-  View
-} from "react-native";
+import { StyleSheet, View } from "react-native";
 
 export default function ScheduleScreen() {
   const { colors } = useTheme();
@@ -32,8 +30,8 @@ function makeStyles(colors: ReturnType<typeof useTheme>["colors"]) {
     container: {
       flex: 1,
       backgroundColor: colors.background,
-      padding: 24,
-      paddingTop: 64,
+      padding: inset.screen,
+      paddingTop: inset.screenTop,
     },
   });
 }
