+++ client/lib/components/ScheduleItemModal.tsx
@@ -4,9 +4,6 @@ import { useTranslation } from "react-i18next";
 import {
   ActivityIndicator,
   Alert,
-  KeyboardAvoidingView,
-  Modal,
-  Platform,
   Pressable,
   ScrollView,
   StyleSheet,
@@ -20,6 +17,8 @@ import { useTheme } from "../bootstrap/ThemeProvider";
 import { Schedule } from "../models/schedule";
 import { inset } from "../theme/spacing";
 import { type } from "../theme/typography";
+import { BottomSheet, makeSheetStyles } from "./BottomSheet";
+import { FormField } from "./FormField";
 
 type ScheduleIconEntry = { name: string; translationKey: string };
 
@@ -35,16 +34,13 @@ export type ScheduleFormData = Omit<Schedule, keyof Models.Document>;
 
 type Props = {
   visible: boolean;
-  /** undefined = add mode, defined = edit mode */
   item?: Schedule;
   nextSortIndex: number;
   onClose: () => void;
   onTimer?: (gameId: string) => void;
-  /** Resolves on success, throws on failure */
   onSave: (data: ScheduleFormData) => Promise<void>;
   onRules?: (gameId: string) => void;
 };
-
 
 function isValidTime(v: string): boolean {
   if (!/^\d{2}:\d{2}$/.test(v)) return false;
@@ -57,54 +53,7 @@ function isValidDuration(v: string): boolean {
   return !isNaN(n) && n > 0;
 }
 
-
-function FormField({
-  icon,
-  label,
-  required,
-  error,
-  children,
-}: {
-  icon: string;
-  label: string;
-  required?: boolean;
-  error?: string;
-  children: React.ReactNode;
-}) {
-  const { colors } = useTheme();
-  const styles = useMemo(() => fieldStyles(colors), [colors]);
-  return (
-    <View style={styles.field}>
-      <View style={styles.labelRow}>
-        <Ionicons name={icon as any} size={13} color={colors.textMuted} />
-        <Text style={styles.label}>
-          {label}
-          {required && <Text style={styles.required}> *</Text>}
-        </Text>
-      </View>
-      {children}
-      {error ? <Text style={styles.error}>{error}</Text> : null}
-    </View>
-  );
-}
-
-function fieldStyles(colors: ReturnType<typeof useTheme>["colors"]) {
-  return StyleSheet.create({
-    field: { gap: 6 },
-    labelRow: { flexDirection: "row", alignItems: "center", gap: 5 },
-    label: { ...type.caption, color: colors.textMuted, textTransform: "uppercase", letterSpacing: 0.5 },
-    required: { color: colors.error },
-    error: { ...type.caption, color: colors.error },
-  });
-}
-
-function IconPicker({
-  value,
-  onChange,
-}: {
-  value: string;
-  onChange: (name: string) => void;
-}) {
+function IconPicker({ value, onChange }: { value: string; onChange: (name: string) => void }) {
   const { colors } = useTheme();
   const styles = useMemo(() => iconPickerStyles(colors), [colors]);
   const { t } = useTranslation(["components"]);
@@ -149,9 +98,9 @@ function iconPickerStyles(colors: ReturnType<typeof useTheme>["colors"]) {
   });
 }
 
-
 export function ScheduleItemModal({ visible, item, nextSortIndex, onClose, onSave, onRules, onTimer }: Props) {
   const { colors } = useTheme();
+  const sheetStyles = useMemo(() => makeSheetStyles(colors), [colors]);
   const styles = useMemo(() => makeStyles(colors), [colors]);
   const { t } = useTranslation(["components"]);
 
@@ -165,7 +114,6 @@ export function ScheduleItemModal({ visible, item, nextSortIndex, onClose, onSav
   const [timeBlurred, setTimeBlurred] = useState(false);
   const [durBlurred, setDurBlurred] = useState(false);
 
-  // Sync state from item when modal opens
   useEffect(() => {
     if (!visible) return;
     if (item) {
@@ -216,239 +164,133 @@ export function ScheduleItemModal({ visible, item, nextSortIndex, onClose, onSav
   }
 
   return (
-    <Modal
+    <BottomSheet
       visible={visible}
-      animationType="slide"
-      transparent
-      onRequestClose={onClose}
-    >
-      <KeyboardAvoidingView
-        style={styles.overlay}
-        behavior={Platform.OS === "ios" ? "padding" : undefined}
-      >
-        <Pressable style={styles.backdrop} onPress={onClose} />
-
-        <View style={styles.sheet}>
-          <View style={styles.sheetHeader}>
-            <Text style={styles.sheetTitle}>
-              {item ? t("schedule.form.editTitle") : t("schedule.form.addTitle")}
-            </Text>
-            <TouchableOpacity style={styles.closeBtn} onPress={onClose}>
-              <Ionicons name="close" size={22} color={colors.textMuted} />
+      onClose={onClose}
+      title={item ? t("schedule.form.editTitle") : t("schedule.form.addTitle")}
+      footer={
+        <>
+          <View style={styles.actionBtns}>
+            <TouchableOpacity
+              style={[styles.actionBtn, !item && styles.actionBtnDisabled]}
+              disabled={!item}
+              onPress={() => item?.gameId && onTimer?.(item.gameId)}
+            >
+              <Ionicons name="timer-outline" size={16} color={!item ? colors.textMuted : colors.text} />
+              <Text style={[styles.actionBtnText, !item && styles.actionBtnTextDisabled]}>
+                {t("schedule.form.actionTimer")}
+              </Text>
+            </TouchableOpacity>
+            <TouchableOpacity
+              style={[styles.actionBtn, (!item?.gameId || !onRules) && styles.actionBtnDisabled]}
+              disabled={!item?.gameId || !onRules}
+              onPress={() => item?.gameId && onRules?.(item.gameId)}
+            >
+              <Ionicons name="list-outline" size={16} color={(!item?.gameId || !onRules) ? colors.textMuted : colors.text} />
+              <Text style={[styles.actionBtnText, (!item?.gameId || !onRules) && styles.actionBtnTextDisabled]}>
+                {t("schedule.form.actionRules")}
+              </Text>
+            </TouchableOpacity>
+            <TouchableOpacity
+              style={[styles.actionBtn, !item && styles.actionBtnDisabled]}
+              disabled={!item}
+            >
+              <Ionicons name="gift-outline" size={16} color={!item ? colors.textMuted : colors.text} />
+              <Text style={[styles.actionBtnText, !item && styles.actionBtnTextDisabled]}>
+                {t("schedule.form.actionLotteries")}
+              </Text>
             </TouchableOpacity>
           </View>
-
-          <ScrollView
-            style={styles.scroll}
-            contentContainerStyle={styles.scrollContent}
-            keyboardShouldPersistTaps="handled"
-            showsVerticalScrollIndicator={false}
+          <Pressable
+            style={[sheetStyles.saveBtn, (!isValid || saving) && sheetStyles.saveBtnDisabled]}
+            onPress={handleSave}
+            disabled={!isValid || saving}
           >
-            <FormField
-              icon="text-outline"
-              label={t("schedule.form.titleField")}
-              required
-            >
-              <TextInput
-                style={styles.input}
-                value={title}
-                onChangeText={setTitle}
-                placeholder={t("schedule.form.titlePlaceholder")}
-                placeholderTextColor={colors.textPlaceholder}
-              />
-            </FormField>
+            {saving
+              ? <ActivityIndicator size="small" color={colors.onAccent} />
+              : <Text style={sheetStyles.saveBtnText}>{t("schedule.form.save")}</Text>
+            }
+          </Pressable>
+        </>
+      }
+    >
+      <FormField icon="text-outline" label={t("schedule.form.titleField")} required>
+        <TextInput
+          style={sheetStyles.input}
+          value={title}
+          onChangeText={setTitle}
+          placeholder={t("schedule.form.titlePlaceholder")}
+          placeholderTextColor={colors.textPlaceholder}
+        />
+      </FormField>
 
-            <FormField
-              icon="shapes-outline"
-              label={t("schedule.form.iconField")}
-              required
-            >
-              <IconPicker value={icon} onChange={setIcon} />
-            </FormField>
+      <FormField icon="shapes-outline" label={t("schedule.form.iconField")} required>
+        <IconPicker value={icon} onChange={setIcon} />
+      </FormField>
 
-            <FormField
-              icon="time-outline"
-              label={t("schedule.form.startTimeField")}
-              required
-              error={timeBlurred && startTime !== "" && !timeValid ? t("schedule.form.validationStartTime") : undefined}
-            >
-              <TextInput
-                style={[styles.input, timeBlurred && startTime !== "" && !timeValid && styles.inputError]}
-                value={startTime}
-                onChangeText={(v) => { setStartTime(v); if (timeBlurred && isValidTime(v)) setTimeBlurred(false); }}
-                onBlur={() => setTimeBlurred(true)}
-                placeholder={t("schedule.form.startTimePlaceholder")}
-                placeholderTextColor={colors.textPlaceholder}
-                keyboardType="numbers-and-punctuation"
-              />
-            </FormField>
+      <FormField
+        icon="time-outline"
+        label={t("schedule.form.startTimeField")}
+        required
+        error={timeBlurred && startTime !== "" && !timeValid ? t("schedule.form.validationStartTime") : undefined}
+      >
+        <TextInput
+          style={[sheetStyles.input, timeBlurred && startTime !== "" && !timeValid && sheetStyles.inputError]}
+          value={startTime}
+          onChangeText={(v) => { setStartTime(v); if (timeBlurred && isValidTime(v)) setTimeBlurred(false); }}
+          onBlur={() => setTimeBlurred(true)}
+          placeholder={t("schedule.form.startTimePlaceholder")}
+          placeholderTextColor={colors.textPlaceholder}
+          keyboardType="numbers-and-punctuation"
+        />
+      </FormField>
 
-            <FormField
-              icon="hourglass-outline"
-              label={t("schedule.form.durationField")}
-              required
-              error={durBlurred && duration !== "" && !durValid ? t("schedule.form.validationDuration") : undefined}
-            >
-              <TextInput
-                style={[styles.input, durBlurred && duration !== "" && !durValid && styles.inputError]}
-                value={duration}
-                onChangeText={(v) => { setDuration(v); if (durBlurred && isValidDuration(v)) setDurBlurred(false); }}
-                onBlur={() => setDurBlurred(true)}
-                placeholder={t("schedule.form.durationPlaceholder")}
-                placeholderTextColor={colors.textPlaceholder}
-                keyboardType="number-pad"
-              />
-            </FormField>
+      <FormField
+        icon="hourglass-outline"
+        label={t("schedule.form.durationField")}
+        required
+        error={durBlurred && duration !== "" && !durValid ? t("schedule.form.validationDuration") : undefined}
+      >
+        <TextInput
+          style={[sheetStyles.input, durBlurred && duration !== "" && !durValid && sheetStyles.inputError]}
+          value={duration}
+          onChangeText={(v) => { setDuration(v); if (durBlurred && isValidDuration(v)) setDurBlurred(false); }}
+          onBlur={() => setDurBlurred(true)}
+          placeholder={t("schedule.form.durationPlaceholder")}
+          placeholderTextColor={colors.textPlaceholder}
+          keyboardType="number-pad"
+        />
+      </FormField>
 
-            <FormField
-              icon="document-text-outline"
-              label={t("schedule.form.descriptionField")}
-            >
-              <TextInput
-                style={[styles.input, styles.inputMultiline]}
-                value={description}
-                onChangeText={setDescription}
-                placeholder={t("schedule.form.descriptionPlaceholder")}
-                placeholderTextColor={colors.textPlaceholder}
-                multiline
-                numberOfLines={3}
-              />
-            </FormField>
+      <FormField icon="document-text-outline" label={t("schedule.form.descriptionField")}>
+        <TextInput
+          style={[sheetStyles.input, sheetStyles.inputMultiline]}
+          value={description}
+          onChangeText={setDescription}
+          placeholder={t("schedule.form.descriptionPlaceholder")}
+          placeholderTextColor={colors.textPlaceholder}
+          multiline
+          numberOfLines={3}
+        />
+      </FormField>
 
-            // TODO: Find a better way to link items to schedule?
-            <FormField
-              icon="game-controller-outline"
-              label={t("schedule.form.gameIdField")}
-            >
-              <TextInput
-                style={styles.input}
-                value={gameId}
-                onChangeText={setGameId}
-                placeholder={t("schedule.form.gameIdPlaceholder")}
-                placeholderTextColor={colors.textPlaceholder}
-                autoCapitalize="none"
-                autoCorrect={false}
-              />
-            </FormField>
-          </ScrollView>
-
-          <View style={styles.footer}>
-            <View style={styles.actionBtns}>
-              <TouchableOpacity
-                style={[styles.actionBtn, !item && styles.actionBtnDisabled]}
-                disabled={!item}
-                onPress={() => item?.gameId && onTimer?.(item.gameId)}
-              >
-                <Ionicons name="timer-outline" size={16} color={!item ? colors.textMuted : colors.text} />
-                <Text style={[styles.actionBtnText, !item && styles.actionBtnTextDisabled]}>
-                  {t("schedule.form.actionTimer")}
-                </Text>
-              </TouchableOpacity>
-              <TouchableOpacity
-                style={[styles.actionBtn, (!item?.gameId || !onRules) && styles.actionBtnDisabled]}
-                disabled={!item?.gameId || !onRules}
-                onPress={() => item?.gameId && onRules?.(item.gameId)}
-              >
-                <Ionicons name="list-outline" size={16} color={(!item?.gameId || !onRules) ? colors.textMuted : colors.text} />
-                <Text style={[styles.actionBtnText, (!item?.gameId || !onRules) && styles.actionBtnTextDisabled]}>
-                  {t("schedule.form.actionRules")}
-                </Text>
-              </TouchableOpacity>
-              <TouchableOpacity
-                style={[styles.actionBtn, !item && styles.actionBtnDisabled]}
-                disabled={!item}
-              >
-                <Ionicons name="gift-outline" size={16} color={!item ? colors.textMuted : colors.text} />
-                <Text style={[styles.actionBtnText, !item && styles.actionBtnTextDisabled]}>
-                  {t("schedule.form.actionLotteries")}
-                </Text>
-              </TouchableOpacity>
-            </View>
-            <Pressable
-              style={[styles.saveBtn, (!isValid || saving) && styles.saveBtnDisabled]}
-              onPress={handleSave}
-              disabled={!isValid || saving}
-            >
-              {saving
-                ? <ActivityIndicator size="small" color="#fff" />
-                : <Text style={styles.saveBtnText}>{t("schedule.form.save")}</Text>
-              }
-            </Pressable>
-          </View>
-        </View>
-      </KeyboardAvoidingView>
-    </Modal>
+      <FormField icon="game-controller-outline" label={t("schedule.form.gameIdField")}>
+        <TextInput
+          style={sheetStyles.input}
+          value={gameId}
+          onChangeText={setGameId}
+          placeholder={t("schedule.form.gameIdPlaceholder")}
+          placeholderTextColor={colors.textPlaceholder}
+          autoCapitalize="none"
+          autoCorrect={false}
+        />
+      </FormField>
+    </BottomSheet>
   );
 }
 
 function makeStyles(colors: ReturnType<typeof useTheme>["colors"]) {
   return StyleSheet.create({
-    overlay: {
-      flex: 1,
-      justifyContent: "flex-end",
-    },
-    backdrop: {
-      ...StyleSheet.absoluteFillObject,
-      backgroundColor: "rgba(0,0,0,0.6)",
-    },
-    sheet: {
-      backgroundColor: colors.surface,
-      borderTopLeftRadius: 20,
-      borderTopRightRadius: 20,
-      borderWidth: 1,
-      borderBottomWidth: 0,
-      borderColor: colors.border,
-      maxHeight: "90%",
-    },
-    sheetHeader: {
-      flexDirection: "row",
-      alignItems: "center",
-      justifyContent: "space-between",
-      paddingHorizontal: inset.card,
-      paddingTop: inset.card,
-      paddingBottom: inset.tight,
-      borderBottomWidth: 1,
-      borderBottomColor: colors.divider,
-    },
-    sheetTitle: {
-      ...type.h3,
-      color: colors.text,
-    },
-    closeBtn: {
-      padding: 6,
-      borderRadius: 8,
-    },
-    scroll: {
-      flexGrow: 0,
-    },
-    scrollContent: {
-      padding: inset.card,
-      gap: inset.group,
-    },
-    input: {
-      ...type.body,
-      color: colors.text,
-      backgroundColor: colors.surfaceHigh,
-      borderWidth: 1,
-      borderColor: colors.border,
-      borderRadius: 10,
-      paddingVertical: 10,
-      paddingHorizontal: inset.card,
-    },
-    inputError: {
-      borderColor: colors.error,
-    },
-    inputMultiline: {
-      minHeight: 72,
-      textAlignVertical: "top",
-      paddingTop: 10,
-    },
-    footer: {
-      padding: inset.card,
-      gap: inset.tight,
-      borderTopWidth: 1,
-      borderTopColor: colors.divider,
-    },
     actionBtns: {
       flexDirection: "row",
       gap: inset.tight,
@@ -476,19 +318,6 @@ function makeStyles(colors: ReturnType<typeof useTheme>["colors"]) {
     },
     actionBtnTextDisabled: {
       color: colors.textMuted,
-    },
-    saveBtn: {
-      backgroundColor: colors.accent,
-      borderRadius: 10,
-      paddingVertical: 14,
-      alignItems: "center",
-    },
-    saveBtnDisabled: {
-      opacity: 0.4,
-    },
-    saveBtnText: {
-      ...type.button,
-      color: colors.text,
     },
   });
 }
